requirements.txt
Django==5.2.4
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-decouple==3.8
drf-spectacular==0.27.2
django-cors-headers==4.4.0
